+++ src/routes/functions/components/FunctionDrawer.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { useEffect, useMemo } from 'react';
 
 import {
     Drawer,
@@ -7,11 +7,9 @@ import {
     useDisclosure,
     DrawerBody,
     VStack,
-    IconButton,
 } from '@chakra-ui/react';
-import { RiDownload2Line } from 'react-icons/ri';
 
-import { downloadFromApi } from '@/api/request';
+import useAuthStore from '@/features/auth/useAuthStore';
 import useUpdateAssetName from '@/features/updateAsset/useUpdateAssetName';
 import { useDocumentTitleEffect } from '@/hooks/useDocumentTitleEffect';
 import useKeyFromPath from '@/hooks/useKeyFromPath';
@@ -19,6 +17,7 @@ import { useSetLocationPreserveParams } from '@/hooks/useLocationWithParams';
 import { PATHS } from '@/paths';
 import DescriptionDrawerSection from '@/routes/functions/components/DescriptionDrawerSection';
 
+import DownloadIconButton from '@/components/DownloadIconButton';
 import DrawerHeader from '@/components/DrawerHeader';
 import {
     DrawerSection,
@@ -47,6 +46,18 @@ const FunctionDrawer = (): JSX.Element => {
         fetchDescription,
         updateFunction,
     } = useFunctionStore();
+    const {
+        info: { organization_id: currentOrganizationId },
+    } = useAuthStore();
+
+    const hasDownloadPermission = useMemo(() => {
+        if (func && currentOrganizationId) {
+            return func.permissions.download.authorized_ids.includes(
+                currentOrganizationId
+            );
+        }
+        return false;
+    }, [func, currentOrganizationId]);
 
     useEffect(() => {
         if (key) {
@@ -74,15 +85,6 @@ const FunctionDrawer = (): JSX.Element => {
         },
         [func?.name]
     );
-
-    const downloadFunction = () => {
-        if (func) {
-            downloadFromApi(
-                func.function.storage_address,
-                `function-${key}.zip`
-            );
-        }
-    };
 
     const { updateNameDialog, updateNameButton } = useUpdateAssetName({
         dialogTitle: 'Rename function',
@@ -112,14 +114,21 @@ const FunctionDrawer = (): JSX.Element => {
                         onClose();
                     }}
                     extraButtons={
-                        <IconButton
-                            aria-label="Download"
-                            variant="ghost"
-                            fontSize="20px"
-                            color="gray.500"
-                            icon={<RiDownload2Line />}
-                            isDisabled={fetchingFunction || !func}
-                            onClick={downloadFunction}
+                        <DownloadIconButton
+                            storageAddress={
+                                func ? func.function.storage_address : ''
+                            }
+                            filename={`function-${key}.zip`}
+                            aria-label={
+                                hasDownloadPermission
+                                    ? 'Download function'
+                                    : "You don't have the download permission for this function"
+                            }
+                            isDisabled={
+                                !hasDownloadPermission ||
+                                fetchingFunction ||
+                                !func
+                            }
                         />
                     }
                     updateNameButton={updateNameButton}